fix : small changes

+++ src/main/java/mate/academy/bookstore/serivce/category/CategoryServiceImpl.java
@@ -56,6 +56,8 @@ public class CategoryServiceImpl implements CategoryService {
 
     @Override
     public void deleteById(Long id) {
-        categoryRepository.deleteById(id);
+        if (categoryRepository.existsById(id)) {
+            categoryRepository.deleteById(id);
+        }
     }
 }
